Migrate fjs/djs/module to .f.mjs
Implementation and proof migrate from authored TypeScript to
JSDoc-typed JavaScript under the stage-1 migration, completing the
fjs/djs cluster (ast, serializer, tokenizer, parser, transpiler, and
now the top-level module all migrated). Drops the type re-export
(`export type { Object, Array, Primitive, Unknown } from './types.ts'`)
that bridged external importers while module.f.ts was still
TypeScript; those importers now go straight to fjs/djs/types.ts.
Updates fjs/types/rtti/validate/proof.f.ts, fjs/types/rtti/parse/proof.f.ts,
fjs/media/json/schema/module.f.ts, and fjs/module.f.ts's import
specifiers, and fixes stale module.f.ts doc references across seven
files — the two type-alias citations already pointed at content moved
out of module.f.ts in an earlier commit of this PR, now corrected to
point at fjs/djs/types.ts.
