+++ fjs/djs/todo/197.md
@@ -9,7 +9,7 @@
 The DJS `Unknown` value shape
 
 ```ts
-// fjs/djs/module.f.ts:25
+// fjs/djs/types.ts:14,16
 type Primitive = JsonPrimitive | bigint | undefined
 type Unknown   = Primitive | Object | Array
 ```
