+++ fjs/types/rtti/parse/proof.f.ts
@@ -3,7 +3,7 @@ import type { ValidationError } from '../common/types.ts'
 import { boolean, number, string, bigint, unknown, array, record, or, option } from '../module.f.mjs'
 import type { Equal } from '../../ts/types.ts'
 import type { Ts } from '../ts/types.ts'
-import type { Unknown as DjsUnknown } from '../../../djs/module.f.ts'
+import type { Unknown as DjsUnknown } from '../../../djs/types.ts'
 import type { Assert } from '../../../asserts/types.ts'
 import { assert, assertEq } from '../../../asserts/module.f.mjs'
