+++ todo/fjs-nanvm-integration.md
@@ -74,7 +74,7 @@ contract and migration strategy.
 ### CLI: an output target, not a command group (decided)
 
 `fjs compile <input> <output>` already dispatches on the output extension
-(`.json` vs. DJS — see [`fjs/djs/module.f.ts`](../fjs/djs/module.f.ts)). Rust
+(`.json` vs. DJS — see [`fjs/djs/module.f.mjs`](../fjs/djs/module.f.mjs)). Rust
 code generation is a third branch, selected by the `.rs` extension:
 
 - `fjs compile <module> <output>.rs` — parse + compile into a generated Rust
